add logic userId

+++ src/services/contacts.js
@@ -3,11 +3,11 @@ import { calculatePaginationData } from "../utils/calculatePaginationData.js";
 import { SORT_ORDER } from "../constants/index.js";
 
 
-export const getAllContacts = async ({ page = 1, perPage = 10, sortBy = '_id', sortOrder = SORT_ORDER.ASC, filter = {} }) => {
+export const getAllContacts = async ({ page = 1, perPage = 10, sortBy = '_id', sortOrder = SORT_ORDER.ASC, filter = {}, userId, }) => {
     const limit = perPage;
     const skip = (page - 1) * perPage;
 
-    const contactsQuery = ContactsCollection.find();
+    const contactsQuery = ContactsCollection.find({userId});
 
     if (filter.type) {
         contactsQuery.where('contactType').equals(filter.type);
@@ -16,6 +16,7 @@ export const getAllContacts = async ({ page = 1, perPage = 10, sortBy = '_id', s
     if (filter.isFavourite !== undefined) {
         contactsQuery.where('isFavourite').equals(filter.isFavourite);
     }
+
 
     const [contactsCount, contacts] = await Promise.all([ContactsCollection.find().merge(contactsQuery).countDocuments(),
         contactsQuery.skip(skip).limit(limit).sort({ [sortBy]: sortOrder }).exec()]);
@@ -29,8 +30,11 @@ export const getAllContacts = async ({ page = 1, perPage = 10, sortBy = '_id', s
     };
 };
 
-export const getAllContactById = async (contactId) => {
-    const contact = await ContactsCollection.findById(contactId);
+export const getAllContactById = async (contactId, userId) => {
+    const contact = await ContactsCollection.findByOne({
+        _id: contactId,
+        userId,
+    });
     return contact;
 };
 
@@ -39,8 +43,8 @@ export const createContact = async (payload) => {
     return contact;
 };
 
-export const updateContact = async (contactId, payload, options ={}) => {
-    const contact = await ContactsCollection.findByIdAndUpdate( contactId, payload, {
+export const updateContact = async (contactId, userId, payload, options ={}) => {
+    const contact = await ContactsCollection.findByIdAndUpdate({_id: contactId, userId, }, payload, {
         new: true,
         ...options
     });
@@ -53,7 +57,7 @@ export const updateContact = async (contactId, payload, options ={}) => {
 
 };
 
-export const deleteContact = async (contactId) => {
-    const contact = await ContactsCollection.findByIdAndDelete({_id: contactId });
+export const deleteContact = async (contactId, userId) => {
+    const contact = await ContactsCollection.findByIdAndDelete({_id: contactId, userId, });
     return contact;
 };
